pyproject.toml
[tool.poetry]
name = "sebastianbot"
version = "0.1.0"
description = "Azure Functions bot for personal automation"
authors = ["omlnaut"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
azure-functions = "^1.24"
python-telegram-bot = "^22.5"
azure-identity = "^1.25"
azure-keyvault-secrets = "^4.10"
pydantic = {extras = ["email"], version = "^2.12.5"}
google-auth-oauthlib = "^1.2"
google-api-python-client = "^2.187"
beautifulsoup4 = "^4.14"
requests = "^2.32.5"
google-genai = "^1.56.0"
astroid = "^4.0.2"
tenacity = "^9.1.2"
azure-eventgrid = "^4.22.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^7.1"
pre-commit = "^4.0"
poetry-plugin-export = "^1.10.0"

[tool.poetry.group.test.dependencies]
pytest = "^9.0"
pytest-cov = "^7.0"
pytest-archon = "^0.0.7"
azure-mgmt-web = "^10.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.coverage.run]
source = ["sebastian/shared", "sebastian/usecases"]

[tool.pyright]
typeCheckingMode = "strict"
reportPrivateUsage = false
reportMissingTypeStubs = false

[tool.isort]
profile = "black"
